refactor :: move SignupHeader into components/auth/layout/AuthStepLayout
-screen/auth 내부 폴더 구조 변경
-AuthHome을 AuthHomeScreen으로 변경
-SignupHeader를 StepHeader로 변경
-StepHeader, ActionLayout을 auth에서 공통으로 쓰일 하위 Layout으로 분류
