+++ src/screens/auth/signup/AdminSignupScreen.tsx
@@ -1,7 +1,7 @@
 import { useEffect, useRef, useState } from "react"
-import SignupHeader from "./layout/SignupHeader";
+import StepHeader from "../../../components/auth/authLayout/AuthStepLayout/StepHeader";
 import IdStep from "./steps/IdStep";
-import ActionLayout from "./layout/ActionLayout";
+import ActionLayout from "../../../components/auth/authLayout/AuthStepLayout/ActionLayout";
 import KeyboardDismiss from "@/components/common/KeyboardDismiss";
 import PasswordStep from "./steps/PasswordStep";
 import VerificationStep from "./steps/VerificationStep";
@@ -23,7 +23,7 @@ export default function AdminSignupScreen() {
     <KeyboardDismiss>
       <AuthStepLayout>
         <>  
-          <SignupHeader step={step} maxStep={4} onPrev={handlePrev} />
+          <StepHeader step={step} maxStep={4} onPrev={handlePrev} />
           <ActionLayout label={step < 4 ? '다음' : step === 4 ? '완료' : '로그인하러 가기'} onNext={handleNext} isValid={stepValid} step={step}>
             <StepContent step={step} form={form} setForm={setForm} setIsEmail={setIsEmail} updateStepValid={updateStepValid} />
           </ActionLayout>
